+++ src/app/(fullscreen)/trait/_components/step1.tsx
@@ -1,8 +1,7 @@
 "use client";
 
 import { useFormContext, Controller } from "react-hook-form";
-import { StepHeader } from "@/components";
-import TypeList from "../../../_components/type-list";
+import { StepHeader, TypeList } from "@/components";
 import { MOOD } from "@/constants/trait";
 
 interface Step1Props {
