+++ src/app/(fullscreen)/trait/_components/step3.tsx
@@ -1,8 +1,7 @@
 "use client";
 
-import { StepHeader } from "@/components";
+import { StepHeader, TypeList } from "@/components";
 import { CONTENT } from "@/constants/trait";
-import TypeList from "../../../_components/type-list";
 import { useFormContext, Controller } from "react-hook-form";
 
 interface Step3Props {
