+++ src/app/_components/index.ts
@@ -9,3 +9,4 @@ export { default as StepHeader } from "./step-header";
 export { default as Header } from "./header";
 export { default as Fixedlayout } from "./fixedlayout";
 export { default as ToggleTab } from "./toggle-tab";
+export { default as TypeList } from "./type-list";
